pyproject.toml
[project]
name = "projet2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "plotly-express>=0.4.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "streamlit>=1.54.0",
    "asttokens>=3.0.1",
    "duckdb>=1.4.4",
    "ipykernel>=7.2.0",
    "jupyter-client>=8.8.0",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "notebook>=7.5.3",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "plotly-express>=0.4.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "textblob>=0.19.0",
]
